emojify/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
click==6.7                # via flask
flask==0.12.2
google-api-python-client==1.6.2
gunicorn==19.7.1
httplib2==0.10.3          # via google-api-python-client, oauth2client
itsdangerous==0.24        # via flask
jinja2==2.9.6             # via flask
markupsafe==1.0           # via jinja2
oauth2client==4.1.2
olefile==0.44             # via pillow
pillow==4.2.1
pyasn1-modules==0.0.9     # via oauth2client
pyasn1==0.2.3             # via oauth2client, pyasn1-modules, rsa
rsa==3.4.2                # via oauth2client
six==1.10.0               # via google-api-python-client, oauth2client
uritemplate==3.0.0        # via google-api-python-client
werkzeug==0.12.2          # via flask
